EFM23.6: IID over FID, and cleanup
This push corresponds to EFM23.6. In this new version, we use IIDs from
ped and bsid files, as opposed to FIDs, which  could contain duplicates.
This push addresses changes in output - it is possible that there is
more internal work needed to enforce this rule - although it does not
appear to be the case. This push also adds some legacy code cleanup.